Show a countdown-driven progress bar in Live Activities
When a Live Activity is started as a countdown (chronometer + when) without explicit progress/progress_max, render a progress bar driven by the countdown timer on the Lock Screen and Dynamic Island. Explicit progress still takes precedence; with neither, no bar is shown.
