+++ Sources/App/Resources/en.lproj/Localizable.strings
@@ -652,6 +652,7 @@ This server requires a client certificate (mTLS) but the operation was cancelled
 "kiosk.sensors.title" = "Sensors";
 "kiosk.title" = "Kiosk Mode";
 "legacy_actions.disclaimer" = "Legacy iOS Actions are not the recommended way to interact with Home Assistant anymore, please use Scripts, Scenes and Automations directly in your Widgets, Apple Watch and CarPlay.";
+"live_activity.accessibility.progress" = "Progress";
 "live_activity.default_title" = "Home Assistant";
 "live_activity.empty_state" = "No active Live Activities";
 "live_activity.end_all.button" = "End All Activities";
