+++ Sources/Extensions/Widgets/LiveActivity/HALockScreenView.swift
@@ -52,6 +52,8 @@ struct HALockScreenView: View {
                     trackColor: trackColor,
                     height: 10
                 )
+            } else if state.chronometer == true, let end = state.countdownEnd {
+                HAActivityTimerProgressBar(end: end, tint: barColor)
             }
         }
         .padding(.horizontal, DesignSystem.Spaces.two)
@@ -138,46 +140,6 @@ struct HALockScreenView: View {
 
     private var trackColor: Color {
         foreground.opacity(useLightText ? 0.14 : 0.08)
-    }
-}
-
-@available(iOS 17.2, *)
-struct HAActivityProgressBar: View {
-    let fraction: Double
-    let fillColor: Color
-    let trackColor: Color
-    let height: CGFloat
-
-    private var clampedFraction: Double {
-        min(max(fraction, 0), 1)
-    }
-
-    var body: some View {
-        GeometryReader { geometry in
-            let width = clampedFraction == 0 ? 0 : max(geometry.size.width * clampedFraction, height)
-
-            ZStack(alignment: .leading) {
-                Capsule(style: .continuous)
-                    .fill(trackColor)
-
-                Capsule(style: .continuous)
-                    .fill(
-                        LinearGradient(
-                            colors: [
-                                fillColor.opacity(0.9),
-                                fillColor,
-                            ],
-                            startPoint: .leading,
-                            endPoint: .trailing
-                        )
-                    )
-                    .frame(width: width)
-            }
-        }
-        .frame(height: height)
-        .accessibilityElement(children: .ignore)
-        .accessibilityLabel("Progress")
-        .accessibilityValue(Text(HAActivityVisualStyle.accessibilityPercentString(for: clampedFraction)))
     }
 }
